# apis/open-api/canvas/canvas-open-api.openapi.json
{
  "openapi": "3.1.2",
...
  "info": {
    "title": "SDKWork Canvas Open API",
    "version": "0.1.0",
    "description": "Public integration authority skeleton for SDKWork Canvas. Open API exposes Notes business resources through API-key context and does not expose app login/session or Drive storage internals."
  },
  "servers": [
    {
      "url": "/canvas/v3/api",
      "description": "SDKWork Canvas Open API root"
    }
  ],
...
  "paths": {
...
    "/canvas/v3/api/boards/{pageId}/versions/{driveVersionId}/restore": {
      "post": {
        "summary": "Restore a Drive-owned page content version through Open API.",
        "operationId": "pages.versions.restore",
        "tags": [
          "canvas"
        ],
        "security": [
          {
            "ApiKey": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PageIdPath"
          },
          {
            "$ref": "#/components/parameters/DriveVersionIdPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestorePageVersionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Restored page content envelope with the new current Drive version.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SdkWorkListResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ProblemDetailResponse"
          }
        },
        "x-sdkwork-owner": "sdkwork-canvas",
        "x-sdkwork-api-authority": "sdkwork-canvas-open-api",
        "x-sdkwork-domain": "canvas",
        "x-sdkwork-resource": "pages",
        "x-sdkwork-permission": "canvas.open.pages.versions.write",
        "x-sdkwork-tenant-scope": "tenant",
        "x-sdkwork-data-scope": "organization",
        "x-sdkwork-audit-event": "canvas.open.page.versions.restore",
        "x-sdkwork-idempotent": false,
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "open-api"
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "parameters": {
...
      "PageIdPath": {
        "name": "pageId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
      "DriveVersionIdPath": {
        "name": "driveVersionId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Drive-owned node version id exposed through the Notes page facade."
      }
    },
    "responses": {
      "ProblemDetailResponse": {
        "description": "Problem detail error.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        }
      }
    },
    "schemas": {
      "FieldError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "integer",
            "format": "int32",
            "minimum": 40011,
            "maximum": 40099
          }
        }
      },
      "ProblemDetail": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "$ref": "#/components/schemas/SdkWorkPlatformErrorCode"
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        }
      },
      "PageInfo": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "mode"
        ],
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "offset",
              "cursor"
            ]
          },
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "pageSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 200
          },
          "totalItems": {
            "type": "string",
            "pattern": "^[0-9]+$"
          },
          "totalPages": {
            "type": "integer",
            "minimum": 0
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          },
          "hasMore": {
            "type": "boolean"
          }
        }
      },
...
      "RestorePageVersionRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "expectedCurrentDriveVersionId": {
            "type": "string"
          }
        }
      },
      "SdkWorkApiResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "data",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "enum": [
              0
            ],
            "default": 0,
            "minimum": 0,
            "maximum": 0
          },
          "data": {
            "description": "Operation-specific payload typed per response schema."
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          }
        }
      },
...
      "SdkWorkPageData": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "pageInfo"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "pageInfo": {
            "$ref": "#/components/schemas/PageInfo"
          }
        }
      },
...
      "SdkWorkPlatformErrorCode": {
        "type": "integer",
        "format": "int32",
        "minimum": 40001,
        "maximum": 79999,
        "description": "Platform or domain error code per API_SPEC.md §15.3."
      },
...
      "SdkWorkListResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/SdkWorkPageData"
              }
            }
          }
        ]
      },
...
    }
  }
}